Updated System Admin for Company Enterprise
Added required model parameters in java files having job listings, job appointments and accepted jobs data models to update the functionality and fix small bugs in System Admin for Company Enterprise
