+++ NiceJobApp/src/UI/CompanyDepartmentJPanel.java
@@ -1831,18 +1831,18 @@ public class CompanyDepartmentJPanel extends javax.swing.JPanel {
     
     public void createPieChart(){
     
-//        DataConnection data = new DataConnection(conn);
-//        Integer count =-20;
-//        
+        DataConnection data = new DataConnection(conn);
+        Integer count =-20;
+        
 //        HashMap<String,Integer> myInp = data.dataConn();
 //        
 //        for (String myKey:myInp.keySet()){
 //            count+=10;
 //            polarAreaChart1.addItem(new ModelPolarAreaChart(new Color((int) ((Math.random() * (255 - 0)) + 0), (int) ((Math.random() * (255 - 0)) + 0), (int) ((Math.random() * (255 - 0)) + 0)), myKey, myInp.get(myKey)));
 //        }
-//        
-//        polarAreaChart1.start();
- 
-   }
+        
+        polarAreaChart1.start();
+    
+    }
     
 }
