+++ NiceJobApp/src/UI/CompanyInterviewJPanel.java
@@ -1163,6 +1163,16 @@ public class CompanyInterviewJPanel extends javax.swing.JPanel {
 //        
 ////        polarAreaChart1.start();
     
-   
+        DataConnection data = new DataConnection(conn);
+        Integer count =-20;
+        
+//        HashMap<String,Integer> myInp = data.dataConn();
+//        
+        //for (String myKey:myInp.keySet()){
+            count+=10;
+//            polarAreaChart1.addItem(new ModelPolarAreaChart(new Color((int) ((Math.random() * (255 - 0)) + 0), (int) ((Math.random() * (255 - 0)) + 0), (int) ((Math.random() * (255 - 0)) + 0)), myKey, myInp.get(myKey)));
+//        }
+        
+//        polarAreaChart1.start();
     
 }
