+++ NiceJobApp/src/UI/SystemAdminHousingJPanelOld.form
@@ -99,7 +99,7 @@
                                           <Component id="txtLocation" linkSize="3" min="-2" pref="119" max="-2" attributes="0"/>
                                       </Group>
                                   </Group>
-                                  <EmptySpace min="-2" pref="143" max="-2" attributes="0"/>
+                                  <EmptySpace min="-2" pref="164" max="-2" attributes="0"/>
                                   <Component id="btnSave" linkSize="9" min="-2" max="-2" attributes="0"/>
                                   <EmptySpace min="-2" pref="32" max="-2" attributes="0"/>
                                   <Component id="btnUpdate" linkSize="9" min="-2" max="-2" attributes="0"/>
@@ -192,14 +192,14 @@
 
           <Layout class="org.netbeans.modules.form.compat2.layouts.support.JScrollPaneSupportLayout"/>
           <SubComponents>
-            <Component class="javax.swing.JTable" name="tblHousingDetails">
+            <Component class="javax.swing.JTable" name="tblStudentDetails">
               <Properties>
                 <Property name="font" type="java.awt.Font" editor="org.netbeans.beaninfo.editors.FontEditor">
-                  <Font name="Trebuchet MS" size="15" style="1"/>
+                  <Font name="Trebuchet MS" size="18" style="1"/>
                 </Property>
                 <Property name="model" type="javax.swing.table.TableModel" editor="org.netbeans.modules.form.editors2.TableModelEditor">
                   <Table columnCount="4" rowCount="4">
-                    <Column editable="false" title="Housing Name" type="java.lang.Long"/>
+                    <Column editable="false" title="Houing Name" type="java.lang.Long"/>
                     <Column editable="false" title="Username" type="java.lang.String"/>
                     <Column editable="false" title="Password" type="java.lang.String"/>
                     <Column editable="false" title="Location" type="java.lang.String"/>
@@ -379,7 +379,7 @@
               <Color blue="ff" green="ff" red="ff" type="rgb"/>
             </Property>
             <Property name="horizontalAlignment" type="int" value="4"/>
-            <Property name="text" type="java.lang.String" value="Company Name"/>
+            <Property name="text" type="java.lang.String" value="Name"/>
           </Properties>
         </Component>
         <Component class="javax.swing.JLabel" name="jLabel9">
@@ -415,7 +415,7 @@
               <Color blue="ff" green="ff" red="ff" type="rgb"/>
             </Property>
             <Property name="horizontalAlignment" type="int" value="0"/>
-            <Property name="text" type="java.lang.String" value="COMPANY DETAILS"/>
+            <Property name="text" type="java.lang.String" value="HOUSING DETAILS"/>
           </Properties>
         </Component>
         <Component class="javax.swing.JTextField" name="txtUsername">
