+++ NiceJobApp/src/UI/SystemAdminHousingJPanelOld.java
@@ -36,14 +36,6 @@ import javax.swing.JPanel;
 import javax.swing.JSplitPane;
 import javax.swing.JTextField;
 import javax.swing.table.DefaultTableModel;
-import model.AcceptedJobs;
-import model.AcceptedJobsDir;
-import model.Company;
-import model.CompanyDir;
-import model.JobAppointments;
-import model.JobAppointmentsDir;
-import model.RHJobPosting;
-import model.RHJobPostingDir;
 import model.UniCollege;
 import model.UniCollegeDir;
 import model.UniDepartment;
@@ -58,7 +50,7 @@ import model.UniversityDir;
  *
  * @author naini
  */
-public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
+public class SystemAdminHousingJPanelOld extends javax.swing.JPanel {
 
     /**
      * Creates new form Uni_ExamCell_JPanel
@@ -75,29 +67,23 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
     Integer selectedDepartmentId = 0;
     Integer selectedRow = -1;
     
-    CompanyDir companyDir = new CompanyDir();
-    AcceptedJobsDir acptJobDir = new AcceptedJobsDir();
-    JobAppointmentsDir jobAptDir = new JobAppointmentsDir();
-    RHJobPostingDir jobPosting = new RHJobPostingDir();
     Connection conn = null;
     
-    public SystemAdminCompanyJPanel(JSplitPane splitPane, Connection conn) {
+    public SystemAdminHousingJPanelOld(JSplitPane splitPane, Connection conn, UniExamCell selectedUniExamCell) {
         initComponents();
         
         this.splitPane = splitPane;
         this.conn = conn;
         
-        tblCompanyDetails.getTableHeader().setFont( new Font( "Trebuchet MS" , Font.PLAIN, 18 ));
-//        this.selectedUniExamCell = selectedUniExamCell;
-//        lblHeading2.setText(this.selectedUniExamCell.getName());
-//        getAllUniversityData();
+        tblStudentDetails.getTableHeader().setFont( new Font( "Trebuchet MS" , Font.PLAIN, 18 ));
+        this.selectedUniExamCell = selectedUniExamCell;
+        lblHeading2.setText(this.selectedUniExamCell.getName());
+        getAllUniversityData();
         
-//        selectedCollegeId = 1;
-//        getAllStudents();
-        getAllCompanies();
+        selectedCollegeId = 1;
+        getAllStudents();
         clearAllFields();
-        populateCompaniesTable(companyDir);
-//        populateStudentTable(uniStudents);
+        populateStudentTable(uniStudents);
     }
 
     /**
@@ -111,7 +97,7 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
 
         kGradientPanel1 = new keeptoo.KGradientPanel();
         jScrollPane1 = new javax.swing.JScrollPane();
-        tblCompanyDetails = new javax.swing.JTable();
+        tblStudentDetails = new javax.swing.JTable();
         jLabel7 = new javax.swing.JLabel();
         btnAdd = new button.Button();
         btnDelete = new button.Button();
@@ -140,8 +126,8 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
         kGradientPanel1.setkStartColor(new java.awt.Color(255, 51, 51));
         kGradientPanel1.setPreferredSize(new java.awt.Dimension(1005, 700));
 
-        tblCompanyDetails.setFont(new java.awt.Font("Trebuchet MS", 1, 15)); // NOI18N
-        tblCompanyDetails.setModel(new javax.swing.table.DefaultTableModel(
+        tblStudentDetails.setFont(new java.awt.Font("Trebuchet MS", 1, 18)); // NOI18N
+        tblStudentDetails.setModel(new javax.swing.table.DefaultTableModel(
             new Object [][] {
                 {null, null, null, null},
                 {null, null, null, null},
@@ -149,7 +135,7 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
                 {null, null, null, null}
             },
             new String [] {
-                "Company Name", "Username", "Password", "Location"
+                "Houing Name", "Username", "Password", "Location"
             }
         ) {
             Class[] types = new Class [] {
@@ -167,7 +153,7 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
                 return canEdit [columnIndex];
             }
         });
-        jScrollPane1.setViewportView(tblCompanyDetails);
+        jScrollPane1.setViewportView(tblStudentDetails);
 
         jLabel7.setFont(new java.awt.Font("Trebuchet MS", 1, 18)); // NOI18N
         jLabel7.setForeground(new java.awt.Color(255, 255, 255));
@@ -243,7 +229,7 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
         lblStudentName.setFont(new java.awt.Font("Trebuchet MS", 1, 18)); // NOI18N
         lblStudentName.setForeground(new java.awt.Color(255, 255, 255));
         lblStudentName.setHorizontalAlignment(javax.swing.SwingConstants.RIGHT);
-        lblStudentName.setText("Company Name");
+        lblStudentName.setText("Name");
 
         jLabel9.setFont(new java.awt.Font("Trebuchet MS", 1, 18)); // NOI18N
         jLabel9.setForeground(new java.awt.Color(255, 255, 255));
@@ -258,7 +244,7 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
         jLabel14.setFont(new java.awt.Font("Trebuchet MS", 1, 24)); // NOI18N
         jLabel14.setForeground(new java.awt.Color(255, 255, 255));
         jLabel14.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        jLabel14.setText("COMPANY DETAILS");
+        jLabel14.setText("HOUSING DETAILS");
 
         txtUsername.setFont(new java.awt.Font("Trebuchet MS", 0, 18)); // NOI18N
         txtUsername.addActionListener(new java.awt.event.ActionListener() {
@@ -335,7 +321,7 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
                                         .addComponent(jLabel17, javax.swing.GroupLayout.PREFERRED_SIZE, 162, javax.swing.GroupLayout.PREFERRED_SIZE)
                                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                                         .addComponent(txtLocation, javax.swing.GroupLayout.PREFERRED_SIZE, 119, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                                .addGap(143, 143, 143)
+                                .addGap(164, 164, 164)
                                 .addComponent(btnSave)
                                 .addGap(32, 32, 32)
                                 .addComponent(btnUpdate)))
@@ -423,35 +409,29 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
 
     private void btnViewSelectedActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnViewSelectedActionPerformed
         // TODO add your handling code here:
-        selectedRow = tblCompanyDetails.getSelectedRow();
+        selectedRow = tblStudentDetails.getSelectedRow();
         if(selectedRow < 0){
             JOptionPane.showMessageDialog(this, "Please select a row");
             return;
         }else{
-            DefaultTableModel tableModel = (DefaultTableModel) tblCompanyDetails.getModel();
-            Company company = (Company) tableModel.getValueAt(selectedRow, 0);
-            displayCompany(company);
-       
+            DefaultTableModel tableModel = (DefaultTableModel) tblStudentDetails.getModel();
+            UniStudent student = (UniStudent) tableModel.getValueAt(selectedRow, 0);
+            displayStudent(student);
+            selectedStudent = student;
         }
     }//GEN-LAST:event_btnViewSelectedActionPerformed
 
     private void btnDeleteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnDeleteActionPerformed
         // TODO add your handling code here:
-        selectedRow = tblCompanyDetails.getSelectedRow();
+        selectedRow = tblStudentDetails.getSelectedRow();
         if(selectedRow < 0){
             JOptionPane.showMessageDialog(this, "Please select a row");
             return;
         }else{
-            DefaultTableModel tableModel = (DefaultTableModel) tblCompanyDetails.getModel();
-            Company company = (Company) tableModel.getValueAt(selectedRow, 0);
-            getAllAcceptedJobs();
-            getAllJobAppointments();
-            getAllJobListings();
-            ArrayList<AcceptedJobs> acptJobDel = acptJobDir.searchByCompanyId(companyDir.searchByCompanyName(company.getName()).getId());
-            ArrayList<JobAppointments> jobAptDel = jobAptDir.searchByCompanyId(companyDir.searchByCompanyName(company.getName()).getId());
-            ArrayList<RHJobPosting> jobListingDel = jobPosting.searchByCompanyName(company.getName());
-            deleteCompany(company,acptJobDel,jobAptDel,jobListingDel);
-            
+            DefaultTableModel tableModel = (DefaultTableModel) tblStudentDetails.getModel();
+            UniStudent student = (UniStudent) tableModel.getValueAt(selectedRow, 0);
+            deleteStudent(student);
+            selectedStudent = student;
         }
     }//GEN-LAST:event_btnDeleteActionPerformed
 
@@ -480,7 +460,7 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
 
     private void btnRefreshTableActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnRefreshTableActionPerformed
         // TODO add your handling code here:
-        populateCompaniesTable(companyDir);
+        populateStudentTable(uniStudents);
     }//GEN-LAST:event_btnRefreshTableActionPerformed
 
     private void btnSearchActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnSearchActionPerformed
@@ -489,8 +469,11 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
         panel.add(new JLabel("Please select a field:"));
         DefaultComboBoxModel model = new DefaultComboBoxModel();
         model.addElement("Name");
-        model.addElement("Location");
-        
+        model.addElement("Sevis Id");
+        model.addElement("Intake");
+        model.addElement("Course");
+        model.addElement("Semester");
+        model.addElement("Department");
         JComboBox comboBox = new JComboBox(model);
         panel.add(comboBox);
 
@@ -498,55 +481,115 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
         if(resultField == JOptionPane.OK_OPTION){
             String fieldSelected = comboBox.getSelectedItem().toString();
 
-
+            DefaultComboBoxModel modelIntake = new DefaultComboBoxModel();
             
-            JTextField txtFieldName = new JTextField("");
-            JTextField txtFieldLocation = new JTextField("");
+            LocalDate today = LocalDate.now();
+            Integer year = today.getYear();
+            ArrayList<Integer> listYears = new ArrayList<Integer>();
+            for(int i = year - 2; i < year+1; i++){
+                listYears.add(i);
+            }
+
+            ArrayList<String> terms = new ArrayList<String>();
+            terms.add("Spring");
+            terms.add("Summer");
+            terms.add("Fall");
+
+            ArrayList<String> termsYears = new ArrayList<String>();
+            for(Integer i : listYears){
+                for(String j: terms){
+                    termsYears.add(j + " " + String.valueOf(i));
+                }
+            }
+
+            for(String str : termsYears){
+                modelIntake.addElement(str);
+            }
+            
+            JComboBox comboBoxIntake = new JComboBox(modelIntake);
+            
+            DefaultComboBoxModel depts = new DefaultComboBoxModel();
+            for(UniDepartment dept : uniDepartments.getUniDepartmentList()){
+                if(dept.getCollege().getId() == selectedCollegeId){
+                    depts.addElement(dept.getName());
+                }
+            }
+            JComboBox comboBoxDepts = new JComboBox(depts);
+
+            DefaultComboBoxModel sems = new DefaultComboBoxModel();
+            for(Integer i = 1; i < 5; i++){
+                sems.addElement(i.toString());
+            }
+            JComboBox comboBoxSems = new JComboBox(sems);
+            
+            JTextField txtField = new JTextField("");
             
             panel.add(new JLabel("Value:"));
             
-            if(fieldSelected.equalsIgnoreCase("Name")){
+            if(fieldSelected.equalsIgnoreCase("Intake")){
                 comboBox.setEnabled(false);
-                panel.add(txtFieldName);
-            }else if(fieldSelected.equalsIgnoreCase("Location")){
+                panel.add(comboBoxIntake);
+            }else if(fieldSelected.equalsIgnoreCase("Department")){
                 comboBox.setEnabled(false);
-                panel.add(txtFieldLocation);
+                panel.add(comboBoxDepts);
+            }else if(fieldSelected.equalsIgnoreCase("Semester")){
+                comboBox.setEnabled(false);
+                panel.add(comboBoxSems);
+            }else{
+                comboBox.setEnabled(false);
+                panel.add(txtField);
             }
 
             int resultValue = JOptionPane.showConfirmDialog(null, panel, "Search by Field", JOptionPane.OK_CANCEL_OPTION, JOptionPane.QUESTION_MESSAGE);
             if(resultValue == JOptionPane.OK_OPTION) {
 
-                String fieldValue=null;
-                if(fieldSelected.equalsIgnoreCase("Name")){
-                    fieldValue = txtFieldName.getText();
-                } else if(fieldSelected.equalsIgnoreCase("Location")){
-                    fieldValue = txtFieldLocation.getText();
+                String fieldValue;
+                if(fieldSelected.equalsIgnoreCase("Intake")){
+                    fieldValue = comboBoxIntake.getSelectedItem().toString();
+                } else if(fieldSelected.equalsIgnoreCase("Department")){
+                    fieldValue = comboBoxDepts.getSelectedItem().toString();
+                } else if(fieldSelected.equalsIgnoreCase("Semester")){
+                    fieldValue = comboBoxSems.getSelectedItem().toString();
+                } else {
+                    fieldValue = txtField.getText();
                 }
 
                 if(fieldValue.equalsIgnoreCase("")){
                     JOptionPane.showMessageDialog(this, "Please enter a value");
+                }else if(fieldSelected.equalsIgnoreCase("Sevis Id") && !isSevisIdValid(fieldValue)){
+                    JOptionPane.showMessageDialog(this, "Please enter a valid Sevis Id");
                 }else if(fieldSelected.equalsIgnoreCase("Name") && !isNameValid(fieldValue)){
                     JOptionPane.showMessageDialog(this, "Please enter a valid Name");
-                }else if(fieldSelected.equalsIgnoreCase("Location") && !isNameValid(fieldValue)){
-                    JOptionPane.showMessageDialog(this, "Please enter a valid Company Name");
+                }else if(fieldSelected.equalsIgnoreCase("Course") && !isCourseNameValid(fieldValue)){
+                    JOptionPane.showMessageDialog(this, "Please enter a valid Course Name");
                 }else{
-                    CompanyDir resultStudentDir = new CompanyDir();
-                    ArrayList<Company> resultStudents = new ArrayList<Company>();
+                    UniStudentDir resultStudentDir = new UniStudentDir();
+                    ArrayList<UniStudent> resultStudents = new ArrayList<UniStudent>();
                     
                     if(fieldSelected.equalsIgnoreCase("Name")){
-                        resultStudents = companyDir.searchByName(fieldValue);
+                        resultStudents = uniStudents.searchByName(fieldValue);
+                    }else if(fieldSelected.equalsIgnoreCase("Sevis Id")){
+                        resultStudents.add(uniStudents.searchBySevisId(fieldValue));
+                    }else if(fieldSelected.equalsIgnoreCase("Intake")){
+                        resultStudents = uniStudents.searchByIntake(fieldValue);
+                    }else if(fieldSelected.equalsIgnoreCase("Course")){
+                        resultStudents = uniStudents.searchByCourse(fieldValue);
+                    }else if(fieldSelected.equalsIgnoreCase("Semester")){
+                        resultStudents = uniStudents.searchBySemester(Integer.valueOf(fieldValue));
+                    }else if(fieldSelected.equalsIgnoreCase("Department")){
+                        UniDepartment department = uniDepartments.searchByName(fieldValue);
+                        resultStudents = uniStudents.searchByDepartment(department);
                     }else{
-                        resultStudents = companyDir.searchByCountry(fieldValue);
+                        resultStudents = null;
                     }
 
                     if(resultStudents.isEmpty()){
                         JOptionPane.showMessageDialog(this, "No Results Found");
-                        populateCompaniesTable(companyDir);
+                        populateStudentTable(uniStudents);
                     }else{
                         JOptionPane.showMessageDialog(this, "Entries found");
-                        
-                        resultStudentDir.setCompanyList(resultStudents);
-                        populateCompaniesTable(resultStudentDir);
+                        resultStudentDir.setUniStudentList(resultStudents);
+                        populateStudentTable(resultStudentDir);
                     }
                 }
             }else{
@@ -559,15 +602,15 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
 
     private void btnUpdateActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnUpdateActionPerformed
         // TODO add your handling code here:
-        DefaultTableModel tableModel = (DefaultTableModel) tblCompanyDetails.getModel();
-        Company company = (Company) tableModel.getValueAt(selectedRow, 0);
-        updateCompany(company);
+        DefaultTableModel tableModel = (DefaultTableModel) tblStudentDetails.getModel();
+        UniStudent student = (UniStudent) tableModel.getValueAt(selectedRow, 0);
+        updateStudent(student);
     }//GEN-LAST:event_btnUpdateActionPerformed
 
     private void btnSaveActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnSaveActionPerformed
         // TODO add your handling code here:
-        
-        saveCompanyDetails();
+        saveStudentDetails();
+
     }//GEN-LAST:event_btnSaveActionPerformed
 
     private void txtUsernameActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txtUsernameActionPerformed
@@ -594,7 +637,7 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
     private javax.swing.JLabel lblHeading1;
     private javax.swing.JLabel lblHeading2;
     private javax.swing.JLabel lblStudentName;
-    private javax.swing.JTable tblCompanyDetails;
+    private javax.swing.JTable tblStudentDetails;
     private javax.swing.JTextField txtCompanyName;
     private javax.swing.JTextField txtLocation;
     private javax.swing.JTextField txtPassword;
@@ -674,6 +717,8 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
         txtUsername.setText("");
         txtLocation.setText("");
         
+        populateCmbBoxDepartments(selectedCollegeId);
+        populateCmbBoxIntake();
         
         
         btnSave.setEnabled(false);
@@ -773,36 +818,7 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
         Matcher matcher = pattern.matcher(str);
         return matcher.matches();
     }
-    public void saveCompanyDetails(){
     
-        if( txtCompanyName.getText().equalsIgnoreCase("") || txtUsername.getText().equalsIgnoreCase("") || 
-                txtPassword.getText().equalsIgnoreCase("") || txtLocation.getText().equalsIgnoreCase("") 
-                ){
-            JOptionPane.showMessageDialog(this, "Please fill all fields");
-        }else{
-            
-                Company comp = companyDir.addCompany();
-                comp.setName(txtCompanyName.getText());
-                comp.setUsername(txtUsername.getText());
-                comp.setPassword(txtPassword.getText());
-                comp.setCountry(txtLocation.getText());
-                comp.setDistrict("Mumbai");
-                comp.setPincode("400076");
-                comp.setState("Maharashtra");
-               
-                
-                
-                
-                
-                saveCompanyToDb(comp);
-                
-                clearAllFields();
-                populateCompaniesTable(companyDir);
-        
-            }
-           
-        
-    }
     public void saveStudentDetails(){
         if( txtCompanyName.getText().equalsIgnoreCase("") || txtUsername.getText().equalsIgnoreCase("") || 
                 txtPassword.getText().equalsIgnoreCase("") || txtLocation.getText().equalsIgnoreCase("") 
@@ -863,7 +879,7 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
         try {
             Transport.send(message);
         } catch (MessagingException ex) {
-            Logger.getLogger(SystemAdminCompanyJPanel.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(SystemAdminHousingJPanelOld.class.getName()).log(Level.SEVERE, null, ex);
         }
         System.out.println("Message sent successfully");
    
@@ -882,7 +898,7 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
             return message;
 
         } catch (Exception ex) {
-            java.util.logging.Logger.getLogger(SystemAdminCompanyJPanel.class.getName()).log(Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(SystemAdminHousingJPanelOld.class.getName()).log(Level.SEVERE, null, ex);
         }
         return null;
     }
@@ -910,37 +926,12 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
                 }
                 st.close();
         } catch (SQLException ex) {
-            Logger.getLogger(SystemAdminCompanyJPanel.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-    
-    public void getAllCompanies(){
-        try {
-            String queryStudents = "SELECT * FROM companies";
-            Statement st = conn.createStatement();
-            ResultSet rs = st.executeQuery(queryStudents);                
-                while (rs.next())
-                {
-                    Company comp = companyDir.addCompany();
-                    
-                    comp.setId(rs.getInt("id"));
-                    comp.setName(rs.getString("company_name"));
-                    comp.setCountry(rs.getString("country"));
-                    comp.setState(rs.getString("state"));
-                    comp.setDistrict(rs.getString("district"));
-                    comp.setPincode(rs.getString("pincode"));
-                    comp.setUsername(rs.getString("username"));
-                    comp.setPassword(rs.getString("password"));
-                    
-                }
-                st.close();
-        } catch (SQLException ex) {
-            Logger.getLogger(SystemAdminCompanyJPanel.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(SystemAdminHousingJPanelOld.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
     
     public void populateStudentTable(UniStudentDir uniStudents){
-        DefaultTableModel tableModel = (DefaultTableModel) tblCompanyDetails.getModel();
+        DefaultTableModel tableModel = (DefaultTableModel) tblStudentDetails.getModel();
         tableModel.setRowCount(0);
         
         for(UniStudent student : uniStudents.getUniStudentList()){
@@ -968,7 +959,7 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
             st.executeUpdate(queryNewStudent); 
             st.close();
         } catch (SQLException ex) {
-            Logger.getLogger(SystemAdminCompanyJPanel.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(SystemAdminHousingJPanelOld.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
     
@@ -1053,7 +1044,7 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
             st.executeUpdate(queryNewStudent);   
             st.close();
         } catch (SQLException ex) {
-            Logger.getLogger(SystemAdminCompanyJPanel.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(SystemAdminHousingJPanelOld.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
     
@@ -1078,204 +1069,7 @@ public class SystemAdminCompanyJPanel extends javax.swing.JPanel {
             st.executeUpdate(queryNewStudent);   
             st.close();
         } catch (SQLException ex) {
-            Logger.getLogger(SystemAdminCompanyJPanel.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(SystemAdminHousingJPanelOld.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
-    
-    public void deleteCompany(Company company, ArrayList<AcceptedJobs> acptDel, ArrayList<JobAppointments> jobAptDel, ArrayList<RHJobPosting> jobListingDel){
-        int result = JOptionPane.showConfirmDialog(null, "Are you sure you want to delete this entry?", "Confirm Deletion", JOptionPane.OK_CANCEL_OPTION, JOptionPane.QUESTION_MESSAGE);
-        if(result == JOptionPane.OK_OPTION ){
-            companyDir.deleteCompany(company);
-            deleteCompanyFromDb(company,acptDel,jobAptDel,jobListingDel);
-            clearAllFields();
-            populateCompaniesTable(companyDir);
-        }else{
-            JOptionPane.showMessageDialog(this, "Deletion cancelled");
-        }
-        
-    }
-    
-    public void deleteCompanyFromDb(Company company, ArrayList<AcceptedJobs> acptDel, ArrayList<JobAppointments> jobAptDel, ArrayList<RHJobPosting> jobListingDel){
-        try {
-            String queryDelCompany = "DELETE FROM companies WHERE "
-                    + "id = '" + company.getId()+ "'";
-            Statement st = conn.createStatement();
-            st.executeUpdate(queryDelCompany);   
-            st.close();
-            
-            Statement stAcptDel = conn.createStatement();
-            for(AcceptedJobs acptDelElement:acptDel){
-                String queryDelAcceptedJob = "DELETE FROM accepted_jobs WHERE "
-                        + "company_id = '" + acptDelElement.getCompanyId()+ "'";
-
-                stAcptDel.executeUpdate(queryDelAcceptedJob);
-            }
-            stAcptDel.close();
-            
-            Statement stJobAptDel = conn.createStatement();
-            for(JobAppointments jobAptDelElement:jobAptDel){
-                String queryDelJobAppointment = "DELETE FROM appointments WHERE "
-                        + "company_id = '" + jobAptDelElement.getCompanyId()+ "'";
-
-                stJobAptDel.executeUpdate(queryDelJobAppointment);
-            }
-            stJobAptDel.close();
-            
-            Statement stJobistingDel = conn.createStatement();
-            for(RHJobPosting jobListing:jobListingDel){
-                String queryDelJobListing = "DELETE FROM job_listings WHERE "
-                        + "company_name = '" + jobListing.getCompName()+ "'";
-
-                stJobistingDel.executeUpdate(queryDelJobListing);
-            }
-            stJobistingDel.close();
-            
-       
-            
-        } catch (SQLException ex) {
-            Logger.getLogger(SystemAdminCompanyJPanel.class.getName()).log(Level.SEVERE, null, ex);
-        }
-        
-    }
-    
-    public void populateCompaniesTable(CompanyDir compDir){
-        DefaultTableModel tableModel = (DefaultTableModel) tblCompanyDetails.getModel();
-        tableModel.setRowCount(0);
-        
-        for(Company company : compDir.getCompanyList()){
-            Object row[] = new Object[5];
-            
-            row[0] = company;
-            row[1] = company.getUsername();
-            row[2] = company.getPassword();
-            row[3] = company.getCountry();
-            
-            
-            tableModel.addRow(row);
-        }
-    }
-
-    public void saveCompanyToDb(Company comp){
-        try {
-            String queryNewCompany = "INSERT into companies (company_name,country, state, district, pincode, username, password) VALUES "
-                    + "('" + comp.getName() + "', '" + comp.getCountry()+ "' , '" + comp.getState() + "' "
-                    + ", '" + comp.getDistrict()+ "' , '" + comp.getPincode()+ "' , '" + comp.getUsername() + "' , '" + comp.getPassword()+"' )";
-            Statement st = conn.createStatement();
-            st.executeUpdate(queryNewCompany); 
-            st.close();
-        } catch (SQLException ex) {
-            Logger.getLogger(SystemAdminCompanyJPanel.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-    
-    public void displayCompany(Company company){
-        txtCompanyName.setText(company.getName());
-        
-        txtPassword.setText(company.getPassword());
-        txtUsername.setText(company.getUsername());
-        txtLocation.setText(company.getCountry());
-        
-        
-        btnSave.setEnabled(false);
-        btnUpdate.setEnabled(true);
-        
-        txtUsername.setEnabled(false);
-    }
-    
-    public void updateCompany(Company oldCompany){
-        if( txtCompanyName.getText().equalsIgnoreCase("") || txtUsername.getText().equalsIgnoreCase("") || 
-                txtPassword.getText().equalsIgnoreCase("") || txtLocation.getText().equalsIgnoreCase("")
-                ){
-            JOptionPane.showMessageDialog(this, "Please fill all fields");
-        }else{
-                Company comp = new Company();
-                
-               
-                comp.setName(txtCompanyName.getText());
-                comp.setUsername(txtUsername.getText());
-                comp.setPassword(txtPassword.getText());
-                comp.setCountry(txtLocation.getText());
-                comp.setDistrict("Mumbai");
-                comp.setPincode("400076");
-                comp.setState("Maharashtra");
-
-                
-                companyDir.updateCompany(oldCompany, comp);
-                             
-                updateCompanyToDb(oldCompany, comp);
-                clearAllFields();
-                populateCompaniesTable(companyDir);
-                
-            }
-        }
-    
-    
-    public void updateCompanyToDb(Company oldCompany, Company company){
-        try {
-            String queryUpdateCompany = "UPDATE companies SET company_name = '" + company.getName() + "' , country = '" + company.getCountry() + "'"
-                    + ", state = '" + company.getState()+ "', district = '" + company.getDistrict()+ "', pincode = '" + company.getPincode()+ "'"
-                    + ", username = '" + company.getUsername() + "', password = '" + company.getPassword() + "' WHERE "
-                    + "company_name = '" + oldCompany.getName()+ "'";
-            Statement st = conn.createStatement();
-            st.executeUpdate(queryUpdateCompany);   
-            st.close();
-        } catch (SQLException ex) {
-            Logger.getLogger(SystemAdminCompanyJPanel.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-    
-    public void getAllAcceptedJobs(){
-        try {
-            String queryStudents = "SELECT * FROM accepted_jobs";
-            Statement st = conn.createStatement();
-            ResultSet rs = st.executeQuery(queryStudents);                
-                while (rs.next())
-                {
-                    AcceptedJobs acptJobs = acptJobDir.addAcceptedJobs();
-                    
-                    acptJobs.setCompanyId(rs.getInt("company_id"));
-                    
-                }
-                st.close();
-        } catch (SQLException ex) {
-            Logger.getLogger(SystemAdminCompanyJPanel.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-    
-    public void getAllJobAppointments(){
-        try {
-            String queryStudents = "SELECT * FROM appointments";
-            Statement st = conn.createStatement();
-            ResultSet rs = st.executeQuery(queryStudents);                
-                while (rs.next())
-                {
-                    JobAppointments jobApt = jobAptDir.addJobAppointment();
-                    
-                    jobApt.setCompanyId(rs.getInt("company_id"));
-                    
-                }
-                st.close();
-        } catch (SQLException ex) {
-            Logger.getLogger(SystemAdminCompanyJPanel.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-    
-    public void getAllJobListings(){
-        try {
-            String queryStudents = "SELECT * FROM job_listings";
-            Statement st = conn.createStatement();
-            ResultSet rs = st.executeQuery(queryStudents);                
-                while (rs.next())
-                {
-                    RHJobPosting jobListing = jobPosting.addRHJobPosting();
-                    
-                    jobListing.setCompName(rs.getString("company_name"));
-                    
-                }
-                st.close();
-        } catch (SQLException ex) {
-            Logger.getLogger(SystemAdminCompanyJPanel.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-    
 }
